requirements.txt
alembic==1.10.3
click==8.1.3
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
